cumulus_client_consensus_aura
Module collator

The core collator logic for Aura - slot claiming, block proposing, and collation packaging.

The Collator struct exposed here is meant to be a component of higher-level logic which actually manages the control flow of the collator - which slots to claim, how many collations to build, when to work, etc.

This module also exposes some standalone functions for common operations when building aura-based collators.

Structs

Collator

A utility struct for writing collation logic that makes use of Aura entirely or in part. See module docs for more details.

Params

Parameters for instantiating a Collator.

SlotClaim

A claim on an Aura slot.

Functions

claim_slot

Attempt to claim a slot derived from the given relay-parent header’s slot.

seal

Seal a block with a signature in the header.