Module cumulus_client_consensus_aura::collators::lookahead

A collator for Aura that looks ahead of the most recently included parachain block when determining what to build upon.

This collator also builds additional blocks when the maximum backlog is not saturated. The size of the backlog is determined by invoking a runtime API. If that runtime API is not supported, this assumes a maximum backlog size of 1.

This takes more advantage of asynchronous backing, though not complete advantage. When the backlog is not saturated, this approach lets the backlog temporarily ‘catch up’ with periods of higher throughput. When the backlog is saturated, we typically fall back to the limited cadence of a single parachain block per relay-chain block.

Despite this, the fact that there is a backlog at all allows us to spend more time building the block, as there is some buffer before it can get posted to the relay-chain. The main limitation is block propagation time - i.e. the new blocks created by an author must be propagated to the next author before their turn.

Structs

Params

Parameters for run.

ParamsWithExport

Parameters for run_with_export.

Functions

run

Run async-backing-friendly Aura.

run_with_export

Run async-backing-friendly Aura.