Function cumulus_client_consensus_aura::collators::lookahead::run_with_export

pub fn run_with_export<Block, P, BI, CIDP, Client, Backend, RClient, CHP, Proposer, CS>(
    _: ParamsWithExport<BI, CIDP, Client, Backend, RClient, CHP, Proposer, CS>,
) -> impl Future<Output = ()> + Send + 'static
where
    Block: BlockT,
    Client: ProvideRuntimeApi<Block> + BlockOf + AuxStore + HeaderBackend<Block> + BlockBackend<Block> + Send + Sync + 'static,
    Client::Api: AuraApi<Block, P::Public> + CollectCollationInfo<Block> + AuraUnincludedSegmentApi<Block>,
    Backend: Backend<Block> + 'static,
    RClient: RelayChainInterface + Clone + 'static,
    CIDP: CreateInherentDataProviders<Block, ()> + 'static,
    CIDP::InherentDataProviders: Send,
    BI: BlockImport<Block> + ParachainBlockImportMarker + Send + Sync + 'static,
    Proposer: ProposerInterface<Block> + Send + Sync + 'static,
    CS: CollatorServiceInterface<Block> + Send + Sync + 'static,
    CHP: ValidationCodeHashProvider<Block::Hash> + Send + 'static,
    P: Pair,
    P::Public: AppPublic + Member + Codec,
    P::Signature: TryFrom<Vec<u8>> + Member + Codec,

Run async-backing-friendly Aura.

This is exactly the same as run, but it supports the optional export of each produced POV to the file system.