Function cumulus_client_consensus_aura::equivocation_import_queue::fully_verifying_import_queue

pub fn fully_verifying_import_queue<P, Client, Block: BlockT, I, CIDP>(
    client: Arc<Client>,
    block_import: I,
    create_inherent_data_providers: CIDP,
    slot_duration: SlotDuration,
    spawner: &impl SpawnEssentialNamed,
    registry: Option<&Registry>,
    telemetry: Option<TelemetryHandle>
) -> BasicQueue<Block>
where
    P: Pair + 'static,
    P::Signature: Codec,
    P::Public: Codec + Debug,
    I: BlockImport<Block, Error = Error> + ParachainBlockImportMarker + Send + Sync + 'static,
    Client: ProvideRuntimeApi<Block> + Send + Sync + 'static,
    <Client as ProvideRuntimeApi<Block>>::Api: BlockBuilderApi<Block> + AuraApi<Block, P::Public>,
    CIDP: CreateInherentDataProviders<Block, ()> + 'static,

Start an import queue for a Cumulus node which checks blocks’ seals and inherent data.

Pass in only inherent data providers which don’t include aura or parachain consensus inherents, e.g. things like timestamp and custom inherents for the runtime.

The others are generated explicitly internally.

This should only be used for runtimes where the runtime does not check all inherents and seals in execute_block (see https://github.com/paritytech/cumulus/issues/2436)