Crate cumulus_client_consensus_aura

The AuRa consensus algoritm for parachains.

This extends the Substrate provided AuRa consensus implementation to make it compatible for parachains. The main entry points for of this consensus algorithm are build_aura_consensus and [import_queue].

For more information about AuRa, the Substrate crate should be checked.

Structs

AuraConsensus

The implementation of the AURA consensus for parachains.

AuraVerifier

A verifier for Aura blocks.

BuildAuraConsensusParams

Paramaters of build_aura_consensus.

BuildAuraWorkerParams

Parameters of build_aura_worker.

BuildVerifierParams

Parameters of build_verifier.

ImportQueueParams

Parameters of import_queue.

SlotProportion

A unit type wrapper to express the proportion of a slot.

Traits

InherentDataProviderExt

Slot specific extension that the inherent data provider needs to implement.

Functions

build_aura_consensus

Build the AuraConsensus.

build_verifier

Build the AuraVerifier.

import_queue

Start an import queue for the Aura consensus algorithm.

slot_duration

Get type of SlotDuration for Aura.

Type Definitions

SlotDuration

Slot duration type for Aura.