Crate cumulus_client_consensus_aura

The AuRa consensus algorithm for parachains.

This extends the Substrate provided AuRa consensus implementation to make it compatible for parachains. The main entry points for of this consensus algorithm are AuraConsensus::build and import_queue.

For more information about AuRa, the Substrate crate should be checked.

Modules

collator

The core collator logic for Aura - slot claiming, block proposing, and collation packaging.

collators

Stock, pure Aura collators.

equivocation_import_queue

Structs

AuraConsensus

The implementation of the AURA consensus for parachains.

AuraVerifier

A verifier for Aura blocks.

BuildAuraConsensusParamsDeprecated

Parameters of AuraConsensus::build.

BuildAuraWorkerParams

Parameters of build_aura_worker.

BuildVerifierParams

Parameters of build_verifier.

ImportQueueParams

Parameters for import_queue.

SlotProportion

A unit type wrapper to express the proportion of a slot.

Traits

InherentDataProviderExt

Slot specific extension that the inherent data provider needs to implement.

Functions

build_verifier

Build the AuraVerifier.

import_queue

Start an import queue for the Aura consensus algorithm.

slot_duration

Get the slot duration for Aura by reading from a runtime API at the best block’s state.