Trait cumulus_client_consensus_aura::InherentDataProviderExt

pub trait InherentDataProviderExt {
    // Required method
    fn slot(&self) -> Slot;
}

Slot specific extension that the inherent data provider needs to implement.

Required Methods

fn slot(&self) -> Slot

The current slot that will be found in the InherentData.